Mas cambios

+++ proyecto_ctf/src/main/resources/db/migration/V1__Initial.sql
@@ -29,15 +29,15 @@ CREATE TABLE users.permissions (
 
 -- Create user_roles table (many-to-many relationship between users and roles)
 CREATE TABLE users.user_roles (
-    user_id INT REFERENCES users.users(id) ON DELETE CASCADE,  -- Fixed the reference to users table
-    role_id INT REFERENCES users.roles(id) ON DELETE CASCADE,  -- Fixed the reference to roles table
+    user_id BIGINT REFERENCES users.users(id) ON DELETE CASCADE,  -- Fixed the reference to users table
+    role_id BIGINT REFERENCES users.roles(id) ON DELETE CASCADE,  -- Fixed the reference to roles table
     PRIMARY KEY (user_id, role_id)
 );
 
 -- Create role_permissions table (many-to-many relationship between roles and permissions)
 CREATE TABLE users.role_permissions (
-    role_id INT REFERENCES users.roles(id) ON DELETE CASCADE,  -- Fixed the reference to roles table
-    permission_id INT REFERENCES users.permissions(id) ON DELETE CASCADE,  -- Fixed the reference to permissions table
+    role_id BIGINT REFERENCES users.roles(id) ON DELETE CASCADE,  -- Fixed the reference to roles table
+    permission_id BIGINT REFERENCES users.permissions(id) ON DELETE CASCADE,  -- Fixed the reference to permissions table
     PRIMARY KEY (role_id, permission_id)
 );
